test-with-pom - reorganise the pom module
