+++ e2e/_lib/pom/core/pom-component.ts
@@ -1,0 +1,7 @@
+import { PomObject } from "./pom-object";
+
+import { Locator } from "@playwright/test";
+
+export abstract class PomComponent extends PomObject {
+  abstract locator(): Locator;
+}
