+++ e2e/_lib/pom/layout/footer-component.ts
@@ -1,5 +1,5 @@
 import { Locator } from "@playwright/test";
-import { PomComponent } from "./pom-core";
+import { PomComponent } from "../core/pom-component";
 
 export class FooterComponent extends PomComponent {
   locator(): Locator {
