+++ e2e/_lib/pom/layout/sidebar-component.ts
@@ -1,6 +1,6 @@
 import { Locator } from "@playwright/test";
-import { PomComponent } from "./pom-core";
-import { UserMyAccountPage } from "./user-my-account-page";
+import { PomComponent } from "../core/pom-component";
+import { UserMyAccountPage } from "../pages/user-my-account-page";
 
 export class SidebarComponent extends PomComponent {
   locator(): Locator {
