+++ e2e/_lib/pom/pages/user-my-account-page.ts
@@ -1,7 +1,7 @@
-import { PawtalPage } from "./pawtal-page";
 import { RoutePath } from "@/lib/route-path";
+import { PomPage } from "../core/pom-page";
 
-export class UserMyAccountPage extends PawtalPage {
+export class UserMyAccountPage extends PomPage {
   public url(): string {
     return this.website().urlOf(RoutePath.USER_MY_ACCOUNT_PAGE);
   }
