+++ e2e/user/can-view-account.spec.ts
@@ -1,9 +1,9 @@
 import { test, expect } from "@playwright/test";
-import { loginKnownUser } from "../_lib/pom/init";
+import { loginKnownUser, sidebarOf } from "../_lib/pom/init";
 
 test("user can view their account", async ({ page }) => {
   const { knownUser, petsPage } = await loginKnownUser(page);
-  const accountPage = await petsPage.sidebar().gotoMyAccount();
+  const accountPage = await sidebarOf(petsPage).gotoMyAccount();
   const { userName, userEmail, userPhoneNumber, userResidency } = knownUser;
   await expect(accountPage.exactText(userName)).toBeVisible();
   await expect(accountPage.exactText(userEmail)).toBeVisible();
